Add cycle counter to implement a "real-time" clock
Since we can't use wall-clock time as the real time clock, we need to
simulate it using the number of CPU cycles.

The emulator, used the host's real-time clock as its own real-time
clock. We can't do this, so we need to emulate this using the number of
CPU cycles and hardcoding, say, 10MHz as the CPU clock.  Unfortunately,
the emulator did not keep the number of CPU cycles.  Instead, it simply
defined it to be equal to the number of instructions.

This does not work because there is an instruction WFI (Wait for
Interrupt) that basically "halts" the CPU until there is an "external"
interrupt.  Our only "external" interrupt is the hardware timer. But
since no instructions are executed during the wait, the hardware timer
would never trigger.

This is why we needed an actual cycle counter.