Add support for test infrastructure
Enable the HTIF exit command to carry a payload, just like
riscv-qemu does. This payload is available from the tohost
value that is stored in the cpu state. The interpreter
obtains this value and returns it upon exit. So now we have
a way for test programs runninng within the emulator to tell
the emulator that the test succeeded or failed with any
desired status code.

Removed annoying warnings related to __int128 type and
-Wpedantic and the fact we don't check the exit code for
write().

Improved names of configuration parameters when creating a
new virtual machine. "kernel" becomes "boot_image".

Likewise, the run.lua script has been modified to parse a
variety of command-line parameters. There is still much to
be improved here.