Move to C++ compiler.
In preparation for the rewrite of the instruction loop, I just moved all
source files to C++.
One of the first modifications will be replacing the riscv_cup_template.h macro.
If needed, we will use C++ templates.

Changes in the loop should not change the number of cycles taken to run
each test.
The new tests.lua file checks the reported value of mcycle against the
expected, besides testing the payload.