Add arithmetic instructions; access state indirectly.
The interpreter loop will only access the CPU state through
a state_access object.
The idea is that when we want to generate the access log for
the final step in the verification game, all we have to do
is invoke the interpreter with a logged_state_access object
without changing any line of code.
Using templates, there should be no penalty for this indirect
access as all methods will be inlined by the compiler.