Reorganize and simplify code.
The greatest change was to the PMAs.
Now, both memory and device ranges have a driver.
The motivation is the peek method, which is can now be used without distinction.
This greatly simplified the code that deals with the Merkle.
All driver methods now receive a pointer to the corresponding PMA.
The modules implementing mmio drivers deal with their own initialization
to ensure the context is set appropriately.

PMAs are also correctly checked for overlap at creation time.

We will slowly add RAII to the code.
For now, a new module unique-c-ptr.h implements smart pointers for
calloc and fopen, so we don't have to worry about releasing memory or
closing file over all possible flow-control paths.
This also simplified code wherever used.

The Merkle tree has been improved for const-correctness.