Strict range checking on PMAs.
The only externally visible change is that now we require the size of
all PMA ranges to be a multiple of the page size (4096 bytes).
This restriction is carried over to the sizes of backing files for flash
drives.
ROM and RAM images do not have this issue because the size of the
backing file is independent of the size of the allocated memory.
However, with flash drives, we use memory mapping, and simplicity
requires the size of the file and of the mapped memory to be the same.
(We could play tricks mapping the abutting region in the address space
to a padding filled with zeros, but this less portable and more
complicated for little gain in generality.)

For consistency, image-kernel (unecessary) and image-rootfs (necessary) now produce backing files padded with zeros to the next multiple of 4096.

Functions that look for the PMA range associated to a given address now
receive the size of the access about to be performed and report failure
if the access would cross a PMA. It is difficult to decide what it would
even mean to perform such an access, even if the adjacent PMA is valid.
So we forbid it outright.

TLB functions have been simplified and factored.