Make machine a proper class.
The machine class, which will absorb the emulator class in due time, is
now in place.
It contains a machine_state structure, and a merkle_tree.
(The slowly disappearing emulator class now contains only machine and
htif objects).
The emulator_config is now machine_config, and will at some point be
required by the only constructor in the machine class.

Created a new file cartesi-constants.h for Cartesi-specific constants.