Improve constructors.
Make sure we delete copy and move assignment and constructors that we
know we should not be using.

Move initialization of machine-config to the many classes themselves.
This alowed us to eliminate the machine-config.cpp file.

Moved cartesi-constants.h into riscv-constants.h.
Included a new enum with the initial values for all registers so there
are no more magic constants lying around.

Finish converting defines to enums so they don't leak out of our
namespace.