Make sure we use a statically linked libfdt
Different versions of libdft may advertise compability with different
versions of the device tree specification.
We dynamically link to the system's libfdt because an upgrade may change
this version and cause a different value to be written to ROM, thereby
changing the machine state.
In the future, we will compile libfdt to RISC-V and run it inside the
ROM itself, so this is a temporary problem and solution.