Add device id to PMA to help with blockchain implementation.
There is a 4-bit field DID in the istart part of the PMA entry that
tells the device id associated to the PMA range.
This will simplify the blockchain implementation, because it can use
this id to index the appropriate behavior for reads and writes to each
range.
See the techpaper for the explanation.