Fix snapshot/rollback bug in GRPC emulator server
The issue was that whenever there was a "breakreason" in the server,
which happens when issuing rollback or snapshot, the server tried to
communicate it's address again to the core manager server.

Change the way the server communicates to core manager server by creating and
destroying the connection object every time a communicate address call
is issued (at the time it happens only once, but it might be used in the
future to signal the core manager server that the server is ready after
executing a snapshot or a rollback).