Fix segfault in sd due to overflow in pma.
The code that searched for the PMA corresponding to a physical memory
address did not take into account the possibility of arithmetic overflow
when adding the physical address to the size of the access. This has now
been fixed.

The other change in this commit is unrelated, and simply allows us
to dump the device tree to a file.