Optimize repeated generation of Merkle trees.
Each memory PMA now keeps a list of dirty pages (i.e., pages that have
been written to). When the update_merkle_tree function is invoked,
it loops over the pages of each PMA and asks if it is dirty. Only the
dirty pages have their node updated. This basically elimintates the cost
of updating the Merkle tree when not much has changed.

There is still much room for improvement.

1) We should have have a means of loading the precomputed hashes for
each PMA from an auxiliary file. These would take only 32/4096 = 0.8% of
the size of the subintended PMA range. It would make updating the Merkle
tree for the first time instantaneous.

2) We should use OpenMP to update multiple pages in parallel. This will
require a few changes to the code, to ensure there are no race
conditions when modifying the tree nodes or when dealing with the FIFO
queue. It also requires us to recompile the Lua interpreter with -fopenmp.

3) We could change the Keccak implementation. I have tried the XKCP
implementation and it is about 30% faster than Cryptopp's. The code
seems solid, but the distribution is horrible. We would need to compile
our own.