Fix bug in fast Merkle tree creation.
Pages should be marked dirty when an entry leaves the write TLB, not
when it enters. Duh. Also, the active TLB entries must be marked as
dirty before updating the tree.

Also disabled OpenMP because it might interact with fork() in the GRPC
server. Will implement parallelism using the thread library directly.

Wondering if Carlo deserves chocolate...