Fix strict aliasing rule violations
This commit fixes a bug in the generation of the log for write_memory
operations. In the process, it eliminates another potential source of
undefined behavior that was left from the original emulator
implementation.

In a nutshell, in C and C++, accessing the same memory region through
pointers of different types leads to undefined behavior. The only
guaranteed safe way of accessing memory in this way is with memcpy. If
the code is written carefully, these will be replaced by direct accesses
during optimization, so there is no performance penalty. The implementation
used to access the host memory through pointers to integers of size 1,
2, 4, and 8. This has now been fixed.

In addition, the TLB implementation used some pointer arithmetic to
quickly convert from target physical address to host addresses. In
particular, it converted the pointer to integer, added another integer
with potential overflow assuming modular arithmetic, and converted back
to pointer. We have now fixed this as well.

The performance impact of these changes should be negligible.