Fix optimization bug in update_merkle_tree_page.
Update_merkle_tree_page is a forced update; not subject to the
"dirty page" optimization. So the optimization was removed.
Otherwise, the modifications performed by the logged_state_access
object are not registered in the Merkle tree.

To check for future errors of this kind, a new function was added that
verifies the entire access log. This function is invoked by
machine::step before it returns the access log.