Improve handling of dependencies
Clone grpc from their repository, instead of having it as a submodule.
This allows us to do a shallow clone. But it does not work for the
submodules of grpc itself. Nevhertheless, it clones a lot faster.

Make sure grpc does not use the system's protobuf.

It sucks that luapp is not installed in the system. :(
I guess we just have to add it to the path somehow.
Either way, we can't run "run.lua" or "tests.lua" directly because
the shebang won't work as is. How to fix this?

Install shared libraries for cryptopp and protobuf

For some copy-and-paste reason, on Linux, the make install for GRPC is
does not install libgrpc++.so.1 link, but rather a bogus  libgrpc++.so.6
link. This does not match the soname for the shared library, so we
rename so.6 to so.1 in our makefile as a workaround.