Add proof generation to incremental Merkle tree
Turns out it is possible to generate the proof for the current page in
the incremental Merkle tree generation. This means all operations needed
by the logger can be performed incrementally. It also means that, when
we want to produce the "meta-hash" (the hash of the Merkle tree of all
Machine State root hashes), we will also be able to do that
incrementally, taking only log memory.