Greatly improve test harness
This is a significant improvement to the regression test harness of the
emulator.

The cartesi-machine-tests.lua script has been revamped to play a variety
of tasks.

As before, it can run all binary tests and report if the payload and
mcycle match what was expected. It can now run just a selection of tests
as well.

More importantly, it can generate the root hashes for a series of mcycle
values and save those to files. It can also generate the json log of
the step for a series of mcycle values and save those to
brotli-compressed files files.

When ground-truth are available, the new Makefile targets it compare
these files with the ground-truth and report any regressions.