Fix bug in xPIE when raising exceptions
Exceptions can be handled in M or S mode, depending on whether
deletagion is configured or not. Let that be mode x.
When an exception is raised, the code running is in mode y.
xPIE should be set to the value of xIE when the exception is raised.
It was being set to the value of yIE.