HTIF yield always sets iflags.Y
When htif.yield_progress is set, the emulator will return from a call to run when HTIF receives the yield progress command.
Otherwise, it will plow through as if nothing happened, except for the
iflags.Y bit being set until the next instruction executes.
It's as if resume was immediatelly done (with a payload of 0).

Same for yield.rollup.

The previous setup would be annoying for solidity.