feat(!): add new options to remote machine server
--server-fd option
Allows a parent process to allocate a listening socket before forking
and execing jsonrpc-remote-cartesi-machine, telling the child process to
use that socket to accept connections.  This way, it knows the address
the socket is bound to and that socket is ready to accept connections
before the server is even run. This is important because the parent can
let the OS pick an an ephemeral port and know its value, and because the
fact the socket is ready eliminates any race condition should the parent
try to connect to it before the server is done with its initialization.

--setpgid option causes the server to break away from its parent process
group, by becoming leader of its own new process group whose id is the
child process id. To avoid a race condition, the parent process should
/also/ set the child's process group id to the child process id. One of
the effects of using this option is to put the server into the
background.  Another effect is that the server (and all of its children)
will be in the same process group. This means they all can be signaled
together (i.e., should we want to kill all of them at the same time.)