refactor: shadow pma device becomes memory range
The shadow PMAs never change once the machine is running.
No need to use the more complicated device mechanism, a read-only memory range is fine.
This allows us to use the standard "peek" for memory ranges with the Mekrle tree.
The old peek needed to use the device-state-access to build the shadow from scratch every time.
It used i-device-state-access, which needed to have read_pma_istart/read_pma_ilength.
This in turn required us to add read_pma_istart/read_pma_ilength to i-state-access.
A huge waste of effort for something that can simply be done in machine.cpp once and for all.