refactor: simplify device handling
Remove context from HTIF device. This is important because it was
the only device used in reproducible code that had a context.
This meant there was some gambiarra in the driver code to check if
the context was nullptr.

Remove read_device and write_device from i_state_access. The code
is now in interpret.cpp. This reduces the amount of code we need
when implementing a new state access class.

Simplify the mock pma implementations used by different state access
classes. I think we should merge them into a single one.